@@ -35,26 +35,26 @@
     "@types/react-router": "^5.0.3",
     "@types/react-router-dom": "^4.3.4",
     "@types/react-router-redux": "^5.0.16",
-    "@typescript-eslint/parser": "2.0.0",
-    "@typescript-eslint/eslint-plugin": "2.0.0",
+    "@typescript-eslint/parser": "2.6.0",
+    "@typescript-eslint/eslint-plugin": "2.6.0",
     "@types/webpack-env": "^1.14.0",
     "add": "^2.0.6",
     "assets-webpack-plugin": "^3.9.10",
     "autoprefixer": "^9.6.1",
-    "babel-eslint": "^10.0.2",
+    "babel-eslint": "^10.0.3",
     "bfj": "^7.0.1",
     "commander": "^2.20.0",
     "copy": "^0.3.2",
     "cross-env": "^5.2.0",
     "css-loader": "^3.1.0",
     "dotenv": "^8.0.0",
-    "eslint": "^5.16.0",
-    "eslint-config-prettier": "^4.1.0",
-    "eslint-loader": "^2.1.2",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-prettier": "^3.0.1",
-    "eslint-plugin-react": "^7.12.4",
-    "eslint-plugin-react-hooks": "^1.6.0",
+    "eslint": "^6.6.0",
+    "eslint-config-prettier": "^6.5.0",
+    "eslint-loader": "^3.0.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-react": "^7.16.0",
+    "eslint-plugin-react-hooks": "^2.2.0",
     "file-loader": "^4.1.0",
     "fork-ts-checker-webpack-plugin": "^1.4.3",
     "fs-extra": "^7.0.1",
