@@ -5,7 +5,7 @@ import { isLoadableComponent } from './utils';
 /**
  * This helps us to make sure all the async code is loaded before rendering.
  */
-export async function ensureReady(routes: AsyncRouteProps[], pathname?: string) {
+export async function ensureReady(routes: AsyncRouteProps[], pathname?: string): Promise<void> {
   await Promise.all(
     routes.map((route) => {
       const match = matchPath(pathname || window.location.pathname, route);
