@@ -1,10 +1,10 @@
 import { AsyncRouteableComponent, AsyncRouteComponentType } from './types';
 
 /** @private is the given object a Function? */
-export const isFunction = <T>(obj: T) => 'function' === typeof obj;
+export const isFunction = <T>(obj: T): boolean => 'function' === typeof obj;
 
 /** @private is the given object an Object? */
-export const isObject = <T>(obj: T) => obj !== null && typeof obj === 'object';
+export const isObject = <T>(obj: T): boolean => obj !== null && typeof obj === 'object';
 
 /** @private is the given object/value a promise? */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -22,6 +22,6 @@ export function isLoadableComponent<Props>(
   return (Component as AsyncRouteComponentType<Props>).load !== undefined;
 }
 
-export function getDisplayName<T>(WrappedComponent: React.ComponentType<T>) {
+export function getDisplayName<T>(WrappedComponent: React.ComponentType<T>): string {
   return WrappedComponent.displayName || WrappedComponent.name || 'Component';
 }
