@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import { withRouter, RouteComponentProps } from 'react-router-dom';
 import { Location } from 'history';
 
@@ -9,26 +9,26 @@ export interface ScrollToTopProps<Params = any> extends RouteComponentProps<Para
 }
 
 export class ScrollToTopView extends React.Component<ScrollToTopProps> {
-  focusOnElement = () => {
+  focusOnElement = (): void => {
     const root = document.querySelector('#root') as HTMLElement;
     setTimeout(() => {
       root.focus();
     }, 100);
   };
 
-  handleChange = () => {
+  handleChange = (): void => {
     const { location, changeHandler } = this.props;
 
     changeHandler && changeHandler(location.pathname);
   };
 
-  componentDidMount() {
+  componentDidMount(): void {
     this.handleChange();
 
     this.focusOnElement();
   }
 
-  componentDidUpdate(prevProps: ScrollToTopProps) {
+  componentDidUpdate(prevProps: ScrollToTopProps): void {
     if (this.props.location.hash.length || this.props.location === prevProps.location) {
       return;
     }
@@ -40,7 +40,7 @@ export class ScrollToTopView extends React.Component<ScrollToTopProps> {
     this.focusOnElement();
   }
 
-  render() {
+  render(): ReactNode {
     return <>{this.props.children}</>;
   }
 }
