@@ -1,3 +1,3 @@
-export function getDisplayName<T>(WrappedComponent: React.ComponentType<T>) {
+export function getDisplayName<T>(WrappedComponent: React.ComponentType<T>): string {
   return WrappedComponent.displayName || WrappedComponent.name || 'Component';
 }
